Release 2021.5.0
New in This Release
====================

* Added option to build dispatcher as a static library
* Added ability to build dispatcher under MinGW
* Fixes for tools and samples

For more information on the preview C++/Python APIs and Samples, see
https://software.intel.com/content/www/us/en/develop/articles/onevpl-preview-examples.html

Issues and Limitations
=====================

C++/Python previews
----------------------
* This is a preview release made available to gather feedback for API
  improvements.  Compatibility with future releases is not guaranteed.

Dispatcher
----------
* MFXInit() and MFXInitEx() functions have been deprecated in the 2.0
  oneVPL Specification.  Use MFXLoad() instead.  See the transition
  guide for more details.
* If initialized with MFXInit() or MFXInitEx(), support is limited to
  the 1.x API and will load Intel(R) Media SDK hardware library rather
  than oneVPL hardware library.
* If both Intel(R) Media SDK and oneVPL hardware libraries are
  installed, only oneVPL libraries will be loaded via MFXLoad(). See the
  oneVPL specification for more details.
* The Media SDK runtimes enabling legacy GPU hardware only provide API
  features up to 1.35, even when initialized with MFXLoad.  For more
  details, see [Upgrading from Intel® Media SDK to Intel® oneAPI Video
  Processing
  Library](https://software.intel.com/content/www/us/en/develop/articles/upgrading-from-msdk-to-onevpl.html).
* The libmfx.dll and libmfx.so.2021.1.11 dispatchers are
  deprecated. They are provided for backwards compatibility only and
  will be removed from a future release.  They do not include any
  changes since the 2021.1.1 release.
* Dispatcher support has been extended to include hardware
  implementations, including those that implement API versions lower
  then 2.0. Applications patterned after the 2021.1.1 release samples,
  rather than the 2.0 specification may fail after re-build as the
  previous sample code made the assumption that 2.0 APIs would be
  available. To remedy this please follow the coding patterns in the
  [oneVPL
  spec](https://spec.oneapi.com/versions/latest/elements/oneVPL/source/API_ref/VPL_disp_api_func.html#_CPPv47MFXLoadv)

Tools and samples
-----------------
* Releases installed prior to the 2.4 API update may cause CMAKE build
  failures on Linux due to compatibility issues.  Prior package versions
  were aligned with release version, rather than API.  As a result
  find_package will look for a current version i.e. 2.4 but find a
  larger legacy usage i.e. 2021.2.2 and assume forward compatibility.
* dpccp-blur sample incorrectly states its usage.  The sample example
  commandline should read:

    dpcpp-blur -i in.i420 -w 128 -h 96
    To view: ffplay -f rawvideo -pixel_format bgra -video_size 256x192 -pixel_format yuv420p out.raw

* legacy-vpp sample in hardware mode is not supported on Sky Lake.
* hello-transcode readme incorrectly implies sample runs on hardware.
* decvpp_tool does not support gen12 platforms on Windows in Hardware
  mode.
* Encode B-frame distance setting is inconsistent between CPU and GPU
  implementations. For the CPU implementation 0 means no B-frames, while
  for the GPU implementation 1 means no B-frames.
* Colorspace flags are required when using sample_* tools, if colorspace
  is not set tools will fail.
* Input size should be aligned to 16 bytes for sample_encode.
* The sample_multi_transcode tool included in this release is only
  intended for GPU operation to remain more closely aligned with the
  version of this tool released with Intel(R) Media SDK.
* The sample_multi_transcode tool has a synchronization issue in 1->N
  pipelines. When decode is faster than encoders, the issue can cause
  the free frames consumed faster than the framework released,
  eventually resulting in tool crash when there is no more memory
  available.
* hello-transcode sample only supports the oneVPL CPU implementation in
  this release.
* sample_decode may stop responding if the input is switched between
  multiple resolutions.
* sample_encode occasionally hangs for (CPU) SVT-HEVC encodes when using
  VBR in Ubuntu 18.04.
* sample_encode does not work if the input is scaled or color converted.
* sample_encode, sample_decode, and sample_vpp are not in sync with
  Media SDK's version of these tools.  This will be corrected in future
  releases.
* Windows samples are installed to C:\Program
  Files (x86)\Intel\oneAPI\vpl\latest\examples by default. This should
  not be a writable location, so you will not be able to build samples
  here. To build examples, copy this folder to another location first.
* When configuring systems to run the OpenVINO interop samples in Linux
  gmmlib version conflicts can occur. Debug shows undefined GmmLib
  symbols as the cause of a runtime error.  This can be mitigated by
  using the media stack integrated with OpenVINO installation.
  Uninstall other versions of libva, iHD, and libgmm then select Media
  SDK when installing OpenVINO as described in the [OpenVINO install
  documentation](https://docs.openvinotoolkit.org/latest/openvino_docs_install_guides_installing_openvino_linux.html#install-openvino).

+++ preview/python/binding/mfxstructures.cpp
@@ -28,7 +28,7 @@ struct CArrayWrapper {
         pyclass.def("__getitem__", [](List *self, int i) -> Element & {
             if (i < 0)
                 i += static_cast<int>(self->len);
-            if (i < 0 || i >= self->len)
+            if (i < 0 || i >= static_cast<int>(self->len))
                 throw py::index_error();
             return self->base[i];
         });
@@ -56,7 +56,8 @@ void init_mfxstructures(const py::module &m) {
     py::class_<mfxI16Pair>(m, "mfxI16Pair")
         .def(py::init<>())
         .def(py::init<>([](int32_t v) {
-            return new mfxI16Pair{ (v >> 16), (v & 0x00ffff) };
+            return new mfxI16Pair{ static_cast<int16_t>(v >> 16),
+                                   static_cast<int16_t>(v & 0x00ffff) };
         }))
         .def(py::init<>([](int16_t x, int16_t y) {
             return new mfxI16Pair{ x, y };
