+++ tools/legacy/sample_common/include/time_statistics.h
@@ -82,7 +82,7 @@ public:
                 "%s Total:%.3lfms(%llu smpls),Avg %.3lfms,StdDev:%.3lfms,Min:%.3lfms,Max:%.3lfms\n"),
             prefix,
             (double)totalTime,
-            numMeasurements,
+            (unsigned long long int)numMeasurements,
             (double)GetAvgTime(false),
             (double)GetTimeStdDev(false),
             (double)GetMinTime(false),
